client and server apps get their own AMQP queues
also update documentation, code structure, dependabot dependencies
