+++ proxy-http-client/Cargo.toml
@@ -28,5 +28,5 @@ tokio-stream = { workspace = true }
 tracing = { workspace = true }
 uuid = { workspace = true }
 intersect-ingress-proxy-common = { path = "../shared-deps", version = "0.1.0" }
-reqwest = "0.12.12"
+reqwest = "0.12.14"
 reqwest-eventsource = "0.6.0"
