+++ proxy-http-client/src/configuration.rs
@@ -4,8 +4,9 @@
 /// 3) if using environment variables, see comment in "get_configuration()" as an example of how nesting works
 /// 4) if using ONLY a file variable, this is determined from the APP_CONFIG_FILE environment variable (environment variables have higher precedence)
 /// 5) Additional logic can be found in shared-deps/src/configuration.rs
-use intersect_ingress_proxy_common::configuration::{BrokerSettings, LogLevel};
 use secrecy::SecretString;
+
+use intersect_ingress_proxy_common::configuration::{BrokerSettings, LogLevel};
 
 #[derive(serde::Deserialize, Clone, Debug)]
 pub struct ExternalProxy {
