+++ proxy-http-client/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod configuration;
+pub mod event_source;
 pub mod poster;
